+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.0] - 2024-03-07
+### Added
+- All API requests now include an explicit default timeout
+- `publish()` and `unpublish()` methods for `Image` objects
+### Changed
+- Improved 403 error handling
+
 ## [2.8.1] - 2024-02-01
 ### Fixed
 - `AirApi.images.create()` now logs an error and still returns the `Image` object when an upload attempt fails